Change the advanced find to include a message in the combo box advising users to select or enter a name.  Added the same functionality to ZK, including the popup search menu.
