+++ client/src/org/compiere/apps/ASearch.java
@@ -21,16 +21,11 @@ import java.awt.Cursor;
 import java.awt.Frame;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
-import java.math.BigDecimal;
-import java.sql.Timestamp;
 import java.util.ArrayList;
-import java.util.logging.Level;
-import java.util.regex.Pattern;
 
 import javax.swing.JComponent;
 import javax.swing.JPopupMenu;
 
-import org.adempiere.exceptions.AdempiereException;
 import org.compiere.apps.search.Find;
 import org.compiere.grid.GridController;
 import org.compiere.model.GridField;
@@ -39,12 +34,9 @@ import org.compiere.model.MQuery;
 import org.compiere.model.MUserQuery;
 import org.compiere.print.ReportEngine;
 import org.compiere.print.Viewer;
-import org.compiere.util.CLogger;
-import org.compiere.util.DisplayType;
 import org.compiere.util.Env;
 import org.compiere.util.KeyNamePair;
 import org.compiere.util.Msg;
-import org.compiere.util.ValueNamePair;
 
 /**
  *	Application Search.
@@ -57,8 +49,6 @@ import org.compiere.util.ValueNamePair;
 public class ASearch implements ActionListener
 {
 	
-	private static final String FIELD_SEPARATOR = "<^>";
-	private static final String SEGMENT_SEPARATOR = "<~>";
 	private GridField[]		m_findFields;
 	
 	/**
@@ -119,8 +109,6 @@ public class ASearch implements ActionListener
 	private JPopupMenu 	m_popup = new JPopupMenu("SearchMenu");
 	/**	The Option List					*/
 	private ArrayList<KeyNamePair>	m_list = new ArrayList<KeyNamePair>();
-	/**	Logger			*/
-	private static CLogger log = CLogger.getCLogger(ASearch.class);
 	
 	private int m_onlycurrentdays;
 	private AppsAction m_appsaction;
@@ -142,14 +130,9 @@ public class ASearch implements ActionListener
 	/**
 	 * 	Get the UserQuery for the table.
 	 *  Fill the list and the popup menu
-	 *  @param invoker component to display popup (optional)
-	 * 	@param AD_Tab_ID Tab ID
-	 *  @param AD_Table_ID Table ID
-	 *  @param AD_User_ID User ID
 	 */
 	private void getSearchTargets ()
 	{
-		boolean baseLanguage = Env.isBaseLanguage(Env.getCtx(), "AD_Window"); 
 		MUserQuery[] search = MUserQuery.get(Env.getCtx(), m_AD_Tab_ID);
 		KeyNamePair pp = null;
 		
@@ -160,7 +143,7 @@ public class ASearch implements ActionListener
 		}		
 		else
 		{	
-			pp = pp = new KeyNamePair (0, Msg.getMsg(Env.getCtx(), "Search"));
+			pp = new KeyNamePair (0, Msg.getMsg(Env.getCtx(), "Search"));
 			m_list.add(pp);
 			m_popup.add(pp.toString()).addActionListener(this);
 		}
@@ -180,7 +163,7 @@ public class ASearch implements ActionListener
 		{
 			launchSearch(pp);
 		}
-	}	//	getZoomTargets
+	}	//	getSearchTargets
 
 	
 	/**
@@ -261,7 +244,7 @@ public class ASearch implements ActionListener
 		if(AD_UserQuery_ID  > 0)
 		{
 			userQuery = new MUserQuery(Env.getCtx(),AD_UserQuery_ID, null);
-			m_query = getQuery(userQuery);
+			m_query = userQuery.getQuery(m_AD_Table_ID, m_targetWindowNo, m_findFields);
 			
 			//			Confirmed query
 			if(m_gt != null && m_gc!= null)
@@ -282,150 +265,7 @@ public class ASearch implements ActionListener
 			}
 		}		
 	}
-	
-	private MQuery getQuery(MUserQuery userQuery) 
-	{	
-		int openBrackets = 0;
-		m_query = new MQuery(userQuery.getAD_Table_ID());
-		m_query.addRestriction(Env.parseContext(Env.getCtx(), m_targetWindowNo,"", false));
-		//m_query.addRestriction(Env.parseContext(Env.getCtx(), m_targetWindowNo, m_whereExtended, false));
-		
-		String code = userQuery.getCode();
-		log.fine("Parse user query: " + code);
-		String[] segments = code.split(Pattern.quote(SEGMENT_SEPARATOR),-1);
-
-		
-		for (int i = 0; i < segments.length; i++)
-		{
-			String columnName = null;
-			String operator = null;
-			Object value = null;
-			Object value2 = null;
-			String andor = null;
-			String left_bracket = null;
-			String right_bracket = null;
-			String[] fields = segments[i].split(Pattern.quote(FIELD_SEPARATOR));				
-			for (int j = 0; j < fields.length; j++)
-			{
-				// column
-				if (j == 0 )  
-				{
-					//for (ValueNamePair vnp : columnValueNamePairs)
-					//{
-					//	if (vnp.getValue().equals(fields[j]))
-					//	{
-							columnName = fields[j];
-							continue;
-						//}
-					//}
-				}
-				// operator
-				else if (j == 1)
-				{
-					for (ValueNamePair vnp : MQuery.OPERATORS)
-					{
-						if (vnp.getValue().equals(fields[j]))
-						{
-							operator = vnp.getValue();
-							break;
-						}
-					}
-				}
-				// value
-				else if ( j == 2  && fields[j].length() > 0 )
-				{
-					GridField field = getTargetMField(columnName);
-					value = parseString(field, fields[j]);
-				}
-				// value 2
-				else if ( j == 3 && fields[j].length() > 0 )
-				{
-						GridField field = getTargetMField(columnName);
-						value2 = parseString(field, fields[j]);
-				}
-				// and/or
-				else if (j == 4 && fields[j].length() > 0 )
-				{
-					if ( i != 0 )
-						andor = fields[j];
-				}
-				else if ( j == 5 && fields[j].length() > 0 )
-				{
-					left_bracket = fields[j];
-				}
-				else if ( j == 6 && fields[j].length() > 0 )
-				{
-					right_bracket = fields[j];
-				}
-			}
-
-			//
-			GridField field = getTargetMField(columnName);
-			if (field == null)
-				continue;
-			//boolean isProductCategoryField = isProductCategoryField(field.getAD_Column_ID());
-			String ColumnSQL = field.getColumnSQL(false);
 				
-			String lBrackets = left_bracket;
-			if ( lBrackets != null )
-				openBrackets += lBrackets.length();
-			String rBrackets = right_bracket;
-			if ( rBrackets != null )
-				openBrackets -= rBrackets.length();
-				
-			boolean and = true;
-			if ( i > 0 )
-					and = !"OR".equals(andor);
-				//	Op
-				Object op = operator;
-				if (op == null)
-					continue;
-				
-				if (value == null)
-				{
-					if ( MQuery.OPERATORS[MQuery.EQUAL_INDEX].equals(op) 
-							||  MQuery.OPERATORS[MQuery.NOT_EQUAL_INDEX].equals(op) )
-					{
-						m_query.addRestriction(ColumnSQL, operator, null,
-								columnName, null, and, openBrackets);
-					}
-					else
-					{
-					continue;
-					}
-				}
-				else 
-				{
-				Object parsedValue = parseValue(field, value);
-				if (parsedValue == null)
-					continue;
-				String infoDisplay = value.toString();
-				if (field.isLookup())
-					infoDisplay = field.getLookup().getDisplay(value);
-				else if (field.getDisplayType() == DisplayType.YesNo)
-					infoDisplay = Msg.getMsg(Env.getCtx(), infoDisplay);
-				
-				if (MQuery.OPERATORS[MQuery.BETWEEN_INDEX].equals(op))
-				{
-					if (value2 == null)
-						continue;
-					Object parsedValue2 = parseValue(field, value2);
-					String infoDisplay_to = value2.toString();
-					if (parsedValue2 == null)
-						continue;
-					m_query.addRangeRestriction(ColumnSQL, parsedValue, parsedValue2,
-								columnName, infoDisplay, infoDisplay_to, and, openBrackets);
-				}
-				else
-					m_query.addRestriction(ColumnSQL, operator, parsedValue,
-								columnName, infoDisplay, and, openBrackets);
-				
-				
-				}
-		 }		
-		return m_query;
-	}
-			
 	/**
 	 * get Query
 	 * @return MQuery
@@ -434,144 +274,5 @@ public class ASearch implements ActionListener
 	{
 		return m_query;
 	}
-	
-	/**
-	 * 	Get Target MField
-	 * 	@param columnName column name
-	 * 	@return MField
-	 */
-	public GridField getTargetMField (String columnName)
-	{
-		if (columnName == null)
-			return null;
-		for (int c = 0; c < m_findFields.length; c++)
-		{
-			GridField field = m_findFields[c];
-			if (columnName.equals(field.getColumnName()))
-				return field;
-		}
-		return null;
-	}	//	getTargetMField
-	
-	/**
-	 * 	Parse String
-	 * 	@param field column
-	 * 	@param in value
-	 * 	@return data type corrected value
-	 */
-	private Object parseString(GridField field, String in)
-	{
-		log.log(Level.FINE, "Parse: " +field + ":" + in);
-		if (in == null)
-			return null;
-		int dt = field.getDisplayType();
-		try
-		{
-			//	Return Integer
-			if (dt == DisplayType.Integer
-				|| (DisplayType.isID(dt) && field.getColumnName().endsWith("_ID")))
-			{
-				int i = Integer.parseInt(in);
-				return new Integer(i);
-			}
-			//	Return BigDecimal
-			else if (DisplayType.isNumeric(dt))
-			{
-				return DisplayType.getNumberFormat(dt).parse(in);
-			}
-			//	Return Timestamp
-			else if (DisplayType.isDate(dt))
-			{
-				long time = 0;
-				try
-				{
-					time = DisplayType.getDateFormat_JDBC().parse(in).getTime();
-					return new Timestamp(time);
-				}
-				catch (Exception e)
-				{
-					log.log(Level.SEVERE, in + "(" + in.getClass() + ")" + e);
-					time = DisplayType.getDateFormat(dt).parse(in).getTime();
-				}
-				return new Timestamp(time);
-			}
-			else if (dt == DisplayType.YesNo)
-				return in.equals("Y");
-			else
-				return in;
-		}
-		catch (Exception ex)
-		{
-			log.log(Level.SEVERE, "Object=" + in, ex);
-			return null;
-		}
-
-	}	//	parseValue
-	
-	/**
-	 * 	Parse Value
-	 * 	@param field column
-	 * 	@param in value
-	 * 	@return data type corrected value
-	 */
-	private Object parseValue (GridField field, Object in)
-	{
-		if (in == null)
-			return null;
-		int dt = field.getDisplayType();
-		try
-		{
-			//	Return Integer
-			if (dt == DisplayType.Integer
-				|| (DisplayType.isID(dt) && field.getColumnName().endsWith("_ID")))
-			{
-				if (in instanceof Integer)
-					return in;
-				int i = Integer.parseInt(in.toString());
-				return new Integer(i);
-			}
-			//	Return BigDecimal
-			else if (DisplayType.isNumeric(dt))
-			{
-				if (in instanceof BigDecimal)
-					return in;
-				return DisplayType.getNumberFormat(dt).parse(in.toString());
-			}
-			//	Return Timestamp
-			else if (DisplayType.isDate(dt))
-			{
-				if (in instanceof Timestamp)
-					return in;
-				long time = 0;
-				try
-				{
-					time = DisplayType.getDateFormat_JDBC().parse(in.toString()).getTime();
-					return new Timestamp(time);
-				}
-				catch (Exception e)
-				{
-					log.log(Level.SEVERE, in + "(" + in.getClass() + ")" + e);
-					time = DisplayType.getDateFormat(dt).parse(in.toString()).getTime();
-				}
-				return new Timestamp(time);
-			}
-			//	Return Y/N for Boolean
-			else if (in instanceof Boolean)
-				return ((Boolean)in).booleanValue() ? "Y" : "N";
-		}
-		catch (Exception ex)
-		{
-			log.log(Level.SEVERE, "Object=" + in, ex);
-			String error = ex.getLocalizedMessage();
-			if (error == null || error.length() == 0)
-				error = ex.toString();
-			StringBuffer errMsg = new StringBuffer();
-			errMsg.append(field.getColumnName()).append(" = ").append(in).append(" - ").append(error);
-			//
-			throw new AdempiereException(errMsg.toString());
-			//return null;
-		}
-
-		return in;
-	}	//	parseValue
+		
 }	//	ASearch
