+++ client/src/org/compiere/apps/search/Find.java
@@ -31,6 +31,8 @@ import java.awt.GridBagLayout;
 import java.awt.Insets;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
+import java.awt.event.FocusEvent;
+import java.awt.event.FocusListener;
 import java.awt.event.InputEvent;
 import java.awt.event.KeyEvent;
 import java.awt.event.WindowAdapter;
@@ -45,6 +47,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Hashtable;
 import java.util.Iterator;
+import java.util.List;
 import java.util.Vector;
 import java.util.logging.Level;
 import java.util.regex.Pattern;
@@ -72,6 +75,9 @@ import javax.swing.table.DefaultTableModel;
 import javax.swing.table.TableCellRenderer;
 import javax.swing.table.TableColumn;
 
+import org.adempiere.webui.component.ListCell;
+import org.adempiere.webui.component.ListItem;
+import org.adempiere.webui.component.Listbox;
 import org.compiere.apps.ADialog;
 import org.compiere.apps.AEnv;
 import org.compiere.apps.ConfirmPanel;
@@ -105,6 +111,7 @@ import org.compiere.util.DB;
 import org.compiere.util.DisplayType;
 import org.compiere.util.Env;
 import org.compiere.util.Msg;
+import org.compiere.util.SecureEngine;
 import org.compiere.util.Util;
 import org.compiere.util.ValueNamePair;
 
@@ -159,6 +166,7 @@ public final class Find extends CDialog
 		m_sLast = "** ".concat(Msg.getMsg(Env.getCtx(), "Last Query")).concat(" **");
 		m_sNew = "** ".concat(Msg.getMsg(Env.getCtx(), "New Query")).concat(" **");
 		m_sTipText = "<".concat(Msg.getMsg(Env.getCtx(),"SelectOrEnterQueryName")).concat(">");
+		m_sToolTipText = Msg.getMsg(Env.getCtx(),"SelectOrEnterQueryNameToolTip");
 		//
 		m_query = new MQuery (tableName);
 		m_query.addRestriction(Env.parseContext(Env.getCtx(), m_targetWindowNo, whereExtended, false));
@@ -231,8 +239,8 @@ public final class Find extends CDialog
 	/** Search messages using translation */
 	private String				m_sLast;
 	private String				m_sNew;
-	private String				m_sTipText;
-	
+	private String				m_sTipText;  // Text to display in ComboBoc	
+	private String				m_sToolTipText;  // Tool tip text to display 
 	
 	//
 	private CPanel southPanel = new CPanel();
@@ -374,9 +382,33 @@ public final class Find extends CDialog
 		bIgnore.setMargin(new Insets(2, 2, 2, 2));
 		bIgnore.setToolTipText(Msg.getMsg(Env.getCtx(),"Ignore"));
 		bIgnore.addActionListener(this);
-		fQueryName.setToolTipText (Msg.getMsg(Env.getCtx(),"SelectOrEnterQueryName"));
+		fQueryName.setToolTipText (m_sToolTipText);
 		fQueryName.setEditable(true);
 		fQueryName.addActionListener(this);
+		fQueryName.getEditor().getEditorComponent().addFocusListener(
+				  new FocusListener(){
+				    public void focusGained(FocusEvent arg0) {
+		        		// fQueryName received the focus - delete the tip text so the user can type without 
+		        		// having to delete the tip.
+		    			int index = fQueryName.getSelectedIndex();
+		    			if(index < 0) 
+		    			{
+		    				if (fQueryName.getSelectedItem() == null || fQueryName.getSelectedItem().equals(m_sTipText))
+		    				{
+		                        fQueryName.setSelectedIndex(-1);
+		                        fQueryName.setSelectedItem("");
+		    				}
+		    			}
+				    }
+
+				    public void focusLost(FocusEvent arg0) {
+		        		// fQueryName lost the focus. If the field is blank, replace the tip text.
+						if (fQueryName.getSelectedItem() != null && fQueryName.getSelectedItem().equals(""))
+		    			{
+		                    fQueryName.setSelectedItem(m_sTipText);
+						}
+				    }
+				  });
 		bSave.setIcon(new ImageIcon(org.compiere.Adempiere.class.getResource("images/Save24.gif")));
 		bSave.setMargin(new Insets(2, 2, 2, 2));
 		bSave.setToolTipText(Msg.getMsg(Env.getCtx(),"Save"));
@@ -973,6 +1005,139 @@ public final class Find extends CDialog
 		}
 	}	//	actionPerformed
 
+    /**
+     * Code the query parameters entered in the table into a string that can be saved in the database.
+     * This is the counterpart to {@link #parseUserQuery()}. 
+     * @return a StringBuffer containing the coded query information.
+     */
+	private StringBuffer codeUserQuery() {
+
+		advancedTable.stopEditor(true);
+
+		m_query = new MQuery(m_tableName);
+		m_query.addRestriction(Env.parseContext(Env.getCtx(), m_targetWindowNo, m_whereExtended, false));
+		StringBuffer code = new StringBuffer();
+		int openBrackets = 0;
+		for (int row = 0; row < advancedTable.getRowCount(); row++)
+		{
+			//	Column
+			Object column = advancedTable.getValueAt(row, INDEX_COLUMNNAME);
+			if (column == null)
+				continue;
+			String ColumnName = column instanceof ValueNamePair ? 
+					((ValueNamePair)column).getValue() : column.toString();
+			String infoName = column.toString();
+			//
+			GridField field = getTargetMField(ColumnName);
+			if (field == null)
+				continue;
+			boolean isProductCategoryField = isProductCategoryField(field.getAD_Column_ID());
+			String ColumnSQL = field.getColumnSQL(false);
+			
+			String lBrackets = (String) advancedTable.getValueAt(row, INDEX_LEFTBRACKET);
+			if ( lBrackets != null )
+				openBrackets += lBrackets.length();
+			String rBrackets = (String) advancedTable.getValueAt(row, INDEX_RIGHTBRACKET);
+			if ( rBrackets != null )
+				openBrackets -= rBrackets.length();
+			
+			boolean and = true;
+			if ( row > 0 )
+				and = !"OR".equals((String) advancedTable.getValueAt(row, INDEX_ANDOR));
+			//	Op
+			Object op = advancedTable.getValueAt(row, INDEX_OPERATOR);
+			if (op == null)
+				continue;
+			String Operator = ((ValueNamePair)op).getValue();
+			
+			//	Value	******
+			Object value = advancedTable.getValueAt(row, INDEX_VALUE);
+			if (value == null)
+			{
+				if ( MQuery.OPERATORS[MQuery.EQUAL_INDEX].equals(op) 
+						||  MQuery.OPERATORS[MQuery.NOT_EQUAL_INDEX].equals(op) )
+				{
+					m_query.addRestriction(ColumnSQL, Operator, null,
+							infoName, null, and, openBrackets);
+					
+					if (code.length() > 0)
+						code.append(SEGMENT_SEPARATOR);
+					code.append(ColumnName)
+					.append(FIELD_SEPARATOR)
+					.append(Operator)
+					.append(FIELD_SEPARATOR)
+					.append("")
+					.append(FIELD_SEPARATOR)
+					.append("")
+					.append(FIELD_SEPARATOR)
+					.append( and ? "AND" : "OR")
+					.append(FIELD_SEPARATOR)
+					.append(lBrackets != null ? lBrackets : "")
+					.append(FIELD_SEPARATOR)
+					.append(rBrackets != null ? rBrackets : "");
+				}
+				else
+				{
+				continue;
+				}
+			}
+			else 
+			{
+				Object parsedValue = parseValue(field, value);
+				if (parsedValue == null)
+					continue;
+				String infoDisplay = value.toString();
+				if (field.isLookup())
+					infoDisplay = field.getLookup().getDisplay(value);
+				else if (field.getDisplayType() == DisplayType.YesNo)
+					infoDisplay = Msg.getMsg(Env.getCtx(), infoDisplay);
+				//	Value2	******
+				Object value2 = null;
+				if (MQuery.OPERATORS[MQuery.BETWEEN_INDEX].equals(op))
+				{
+					value2 = advancedTable.getValueAt(row, INDEX_VALUE2);
+					if (value2 == null)
+						continue;
+					Object parsedValue2 = parseValue(field, value2);
+					String infoDisplay_to = value2.toString();
+					if (parsedValue2 == null)
+						continue;
+					m_query.addRangeRestriction(ColumnSQL, parsedValue, parsedValue2,
+								infoName, infoDisplay, infoDisplay_to, and, openBrackets);
+				}
+				else if (isProductCategoryField && MQuery.OPERATORS[MQuery.EQUAL_INDEX].equals(op)) {
+					if (!(parsedValue instanceof Integer)) {
+						continue;
+					}
+					m_query.addRestriction(getSubCategoryWhereClause(((Integer) parsedValue).intValue()), 
+							and, openBrackets);
+				}
+				else
+					m_query.addRestriction(ColumnSQL, Operator, parsedValue,
+								infoName, infoDisplay, and, openBrackets);
+				
+				if (code.length() > 0)
+					code.append(SEGMENT_SEPARATOR);
+				code.append(ColumnName)
+					.append(FIELD_SEPARATOR)
+					.append(Operator)
+					.append(FIELD_SEPARATOR)
+					.append(value.toString())
+					.append(FIELD_SEPARATOR)
+					.append(value2 != null ? value2.toString() : "")
+					.append(FIELD_SEPARATOR)
+					.append( and ? "AND" : "OR")
+					.append(FIELD_SEPARATOR)
+					.append(lBrackets != null ? lBrackets : "")
+					.append(FIELD_SEPARATOR)
+					.append(rBrackets != null ? rBrackets : "");
+				
+			}
+		}
+		
+		return code;
+	}
+
 	/**
 	 * Parse delimited string into user query
 	 * Old field sequence: column, operator, value, value to
@@ -1262,183 +1427,69 @@ public final class Find extends CDialog
 	}	//	cmd_new
 
 	/**
-	 *	Save (Advanced)
+	 * Save the advanced query in the database using the query name. If the query name is set,
+	 * the query will be updated or a new query saved.
+	 * @param saveQuery	Save the query as the Last Query.  Set to true when running the query.
+	 * Set to false to only save using the query name.
 	 */
 	private void cmd_save(boolean saveQuery)
 	{
-		advancedTable.stopEditor(true);
-		//
-		m_query = new MQuery(m_tableName);
-		m_query.addRestriction(Env.parseContext(Env.getCtx(), m_targetWindowNo, m_whereExtended, false));
-		StringBuffer code = new StringBuffer();
-		int openBrackets = 0;
-		for (int row = 0; row < advancedTable.getRowCount(); row++)
-		{
-			//	Column
-			Object column = advancedTable.getValueAt(row, INDEX_COLUMNNAME);
-			if (column == null)
-				continue;
-			String ColumnName = column instanceof ValueNamePair ? 
-					((ValueNamePair)column).getValue() : column.toString();
-			String infoName = column.toString();
-			//
-			GridField field = getTargetMField(ColumnName);
-			if (field == null)
-				continue;
-			boolean isProductCategoryField = isProductCategoryField(field.getAD_Column_ID());
-			String ColumnSQL = field.getColumnSQL(false);
-			
-			String lBrackets = (String) advancedTable.getValueAt(row, INDEX_LEFTBRACKET);
-			if ( lBrackets != null )
-				openBrackets += lBrackets.length();
-			String rBrackets = (String) advancedTable.getValueAt(row, INDEX_RIGHTBRACKET);
-			if ( rBrackets != null )
-				openBrackets -= rBrackets.length();
-			
-			boolean and = true;
-			if ( row > 0 )
-				and = !"OR".equals((String) advancedTable.getValueAt(row, INDEX_ANDOR));
-			//	Op
-			Object op = advancedTable.getValueAt(row, INDEX_OPERATOR);
-			if (op == null)
-				continue;
-			String Operator = ((ValueNamePair)op).getValue();
-			
-			//	Value	******
-			Object value = advancedTable.getValueAt(row, INDEX_VALUE);
-			if (value == null)
-			{
-				if ( MQuery.OPERATORS[MQuery.EQUAL_INDEX].equals(op) 
-						||  MQuery.OPERATORS[MQuery.NOT_EQUAL_INDEX].equals(op) )
-				{
-					m_query.addRestriction(ColumnSQL, Operator, null,
-							infoName, null, and, openBrackets);
-					
-					if (code.length() > 0)
-						code.append(SEGMENT_SEPARATOR);
-					code.append(ColumnName)
-					.append(FIELD_SEPARATOR)
-					.append(Operator)
-					.append(FIELD_SEPARATOR)
-					.append("")
-					.append(FIELD_SEPARATOR)
-					.append("")
-					.append(FIELD_SEPARATOR)
-					.append( and ? "AND" : "OR")
-					.append(FIELD_SEPARATOR)
-					.append(lBrackets != null ? lBrackets : "")
-					.append(FIELD_SEPARATOR)
-					.append(rBrackets != null ? rBrackets : "");
-				}
-				else
-				{
-				continue;
-				}
-			}
-			else 
-			{
-				Object parsedValue = parseValue(field, value);
-				if (parsedValue == null)
-					continue;
-				String infoDisplay = value.toString();
-				if (field.isLookup())
-					infoDisplay = field.getLookup().getDisplay(value);
-				else if (field.getDisplayType() == DisplayType.YesNo)
-					infoDisplay = Msg.getMsg(Env.getCtx(), infoDisplay);
-				//	Value2	******
-				Object value2 = null;
-				if (MQuery.OPERATORS[MQuery.BETWEEN_INDEX].equals(op))
-				{
-					value2 = advancedTable.getValueAt(row, INDEX_VALUE2);
-					if (value2 == null)
-						continue;
-					Object parsedValue2 = parseValue(field, value2);
-					String infoDisplay_to = value2.toString();
-					if (parsedValue2 == null)
-						continue;
-					m_query.addRangeRestriction(ColumnSQL, parsedValue, parsedValue2,
-								infoName, infoDisplay, infoDisplay_to, and, openBrackets);
-				}
-				else if (isProductCategoryField && MQuery.OPERATORS[MQuery.EQUAL_INDEX].equals(op)) {
-					if (!(parsedValue instanceof Integer)) {
-						continue;
-					}
-					m_query.addRestriction(getSubCategoryWhereClause(((Integer) parsedValue).intValue()), 
-							and, openBrackets);
-				}
-				else
-					m_query.addRestriction(ColumnSQL, Operator, parsedValue,
-								infoName, infoDisplay, and, openBrackets);
-				
-				if (code.length() > 0)
-					code.append(SEGMENT_SEPARATOR);
-				code.append(ColumnName)
-					.append(FIELD_SEPARATOR)
-					.append(Operator)
-					.append(FIELD_SEPARATOR)
-					.append(value.toString())
-					.append(FIELD_SEPARATOR)
-					.append(value2 != null ? value2.toString() : "")
-					.append(FIELD_SEPARATOR)
-					.append( and ? "AND" : "OR")
-					.append(FIELD_SEPARATOR)
-					.append(lBrackets != null ? lBrackets : "")
-					.append(FIELD_SEPARATOR)
-					.append(rBrackets != null ? rBrackets : "");
-				
-			}
-		}
-		
-		
-		//  Save the query
-		//  Every query is saved automatically as ** Last Query ** when run.
-		//  Queries run without a name will not be saved.
-
+		//  Update or save the query if there is a valid name.
 		Object selected = fQueryName.getSelectedItem();
 		
-		if (selected != null) {
-			String name = selected.toString();
-			
-			if (name.equals(m_sLast) || name.equals(m_sTipText) || Util.isEmpty(name, true )) 
+		if (selected == null)
+		{
+			return;
+		}
+		
+		//  Generate the query code
+		StringBuffer code = codeUserQuery();
+		
+		String name = selected.toString();
+		
+		if (name.equals(m_sNew) || name.equals(m_sLast) || name.equals(m_sTipText) || Util.isEmpty(name, true )) 
+		{
+			// No name to save to.  Just run the query.
+		}
+		else
+		{
+			MUserQuery uq = MUserQuery.get(Env.getCtx(), m_AD_Tab_ID, name);
+			if (code.length() > 0) { // New or updated
+				if (uq == null) // Create a new record
+				{				
+					uq = new MUserQuery (Env.getCtx(), 0, null);
+					uq.setName (name);
+					uq.setAD_Table_ID(m_AD_Table_ID);
+					uq.setAD_Tab_ID(m_AD_Tab_ID); //red1 UserQuery [ 1798539 ] taking in new field from Compiere
+					uq.setAD_User_ID(Env.getAD_User_ID(Env.getCtx())); //red1 - [ 1798539 ] missing in Compiere delayed source :-)
+				}			
+				uq.setCode (code.toString());  // Update the query code
+				
+			} 
+			else if (code.length() <= 0) // Delete the query
 			{
-				// No name to save to.  Just run the query.
-			}
-			else if (saveQuery)
-			{
-				MUserQuery uq = MUserQuery.get(Env.getCtx(), m_AD_Tab_ID, name);
-				if (code.length() > 0) { // New or updated
-					if (uq == null) // Create a new record
-					{				
-						uq = new MUserQuery (Env.getCtx(), 0, null);
-						uq.setName (name);
-						uq.setAD_Table_ID(m_AD_Table_ID);
-						uq.setAD_Tab_ID(m_AD_Tab_ID); //red1 UserQuery [ 1798539 ] taking in new field from Compiere
-						uq.setAD_User_ID(Env.getAD_User_ID(Env.getCtx())); //red1 - [ 1798539 ] missing in Compiere delayed source :-)
-					}			
-					uq.setCode (code.toString());  // Update the query code
-					
-				} 
-				else if (code.length() <= 0) // Delete the query
+				if (uq.delete(true))
 				{
-					if (uq.delete(true))
-					{
-						ADialog.info (m_targetWindowNo, this, "Deleted", name);
-						refreshUserQueries();
-					}
-					else
-						ADialog.warn (m_targetWindowNo, this, "DeleteError", name);
-					return;
-				}
-				//
-				if (uq.save())
-				{
-					//ADialog.info (m_targetWindowNo, this, "Saved", name);
+					ADialog.info (m_targetWindowNo, this, "Deleted", name);
 					refreshUserQueries();
 				}
 				else
-					ADialog.warn (m_targetWindowNo, this, "SaveError", name);
+					ADialog.warn (m_targetWindowNo, this, "DeleteError", name);
+				return;
 			}
 			//
+			if (uq.save())
+			{
+				//ADialog.info (m_targetWindowNo, this, "Saved", name);
+				refreshUserQueries();
+			}
+			else
+				ADialog.warn (m_targetWindowNo, this, "SaveError", name);
+		}
+		
+		//  Save the query as the Last query run.
+		if (saveQuery)
+		{
 			MUserQuery last = MUserQuery.get(Env.getCtx(), m_AD_Tab_ID, m_sLast);
 			if (code.length() > 0) { // New or update				
 				if (last == null) // Create a new record
@@ -1458,8 +1509,7 @@ public final class Find extends CDialog
 
 			if (!last.save())
 				ADialog.warn (m_targetWindowNo, this, "SaveError", name);
-
-		}
+		}	
 	}	//	cmd_save
 
 	private void refreshUserQueries()
